# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GeminiConversation {
  id: string;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: string;
  conversationId: string;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateGeminiConversationBody {
  title: string;
}

export interface SendGeminiMessageBody {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: string;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiError {
  error: string;
}

export interface MythFact {
  id: number;
  myth: string;
  fact: string;
  category: string;
  source: string;
}

export interface PollingPlace {
  id: number;
  name: string;
  address: string;
  city: string;
  state: string;
  hours: string;
  waitTime: number;
  latitude: number;
  longitude: number;
}

export interface VoterJourneyStep {
  id: number;
  title: string;
  description: string;
  deadline: string | null;
  status: string;
  link: string | null;
  linkText: string | null;
}

export interface VoterJourney {
  state: string;
  steps: VoterJourneyStep[];
  nextElectionDate: string | null;
}

export interface BallotCandidate {
  name: string;
  party: string;
  office: string;
  website: string | null;
}

export interface BallotMeasure {
  title: string;
  description: string;
  type: string;
}

export interface BallotInfo {
  state: string;
  electionDate: string;
  candidates: BallotCandidate[];
  measures: BallotMeasure[];
}

export interface ElectionStats {
  registeredVoters: number;
  voterTurnout2019: number;
  voterTurnoutState: number;
  totalStates: number;
  totalUnionTerritories: number;
  mythsBusted: number;
}

export type SearchPollingPlacesParams = {
  state?: string;
};

export type GetVoterJourneyParams = {
  state?: string;
};

export type GetBallotInfoParams = {
  state: string;
};