# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListGeminiConversationsResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListGeminiConversationsResponse = zod.array(
  ListGeminiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateGeminiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetGeminiConversationParams = zod.object({
  id: zod.coerce.string(),
});

export const GetGeminiConversationResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.string(),
      conversationId: zod.string(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteGeminiConversationParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary List messages in a conversation
 */
export const ListGeminiMessagesParams = zod.object({
  id: zod.coerce.string(),
});

export const ListGeminiMessagesResponseItem = zod.object({
  id: zod.string(),
  conversationId: zod.string(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListGeminiMessagesResponse = zod.array(
  ListGeminiMessagesResponseItem,
);

/**
 * @summary Send a message and receive an AI response (SSE stream)
 */
export const SendGeminiMessageParams = zod.object({
  id: zod.coerce.string(),
});

export const SendGeminiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary List election myths and facts
 */
export const ListMythsResponseItem = zod.object({
  id: zod.number(),
  myth: zod.string(),
  fact: zod.string(),
  category: zod.string(),
  source: zod.string(),
});
export const ListMythsResponse = zod.array(ListMythsResponseItem);

/**
 * @summary Search polling places by state
 */
export const SearchPollingPlacesQueryParams = zod.object({
  state: zod.coerce.string().optional(),
});

export const SearchPollingPlacesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  address: zod.string(),
  city: zod.string(),
  state: zod.string(),
  hours: zod.string(),
  waitTime: zod.number(),
  latitude: zod.number(),
  longitude: zod.number(),
});
export const SearchPollingPlacesResponse = zod.array(
  SearchPollingPlacesResponseItem,
);

/**
 * @summary Get voter journey steps for a state
 */
export const GetVoterJourneyQueryParams = zod.object({
  state: zod.coerce.string().optional(),
});

export const GetVoterJourneyResponse = zod.object({
  state: zod.string(),
  steps: zod.array(
    zod.object({
      id: zod.number(),
      title: zod.string(),
      description: zod.string(),
      deadline: zod.string().nullable(),
      status: zod.string(),
      link: zod.string().nullable(),
      linkText: zod.string().nullable(),
    }),
  ),
  nextElectionDate: zod.string().nullable(),
});

/**
 * @summary Get ballot information by state
 */
export const GetBallotInfoQueryParams = zod.object({
  state: zod.coerce.string(),
});

export const GetBallotInfoResponse = zod.object({
  state: zod.string(),
  electionDate: zod.string(),
  candidates: zod.array(
    zod.object({
      name: zod.string(),
      party: zod.string(),
      office: zod.string(),
      website: zod.string().nullable(),
    }),
  ),
  measures: zod.array(
    zod.object({
      title: zod.string(),
      description: zod.string(),
      type: zod.string(),
    }),
  ),
});

/**
 * @summary Get overall election participation statistics
 */
export const GetElectionStatsResponse = zod.object({
  registeredVoters: zod.number(),
  voterTurnout2019: zod.number(),
  voterTurnoutState: zod.number(),
  totalStates: zod.number(),
  totalUnionTerritories: zod.number(),
  mythsBusted: zod.number(),
});